ホイールズームフックを分離
